requirements.txt
numpy
pandas
pydantic
PyYAML
yfinance
requests
beautifulsoup4
pytest
jinja2
weasyprint
markdown
matplotlib
scipy
